refactor: hoist DiskArtifactCache to fluree-db-core (native)
Move the disk cache machinery (LRU byte budget, eviction, in-flight fetch
coalescing, for_dir singleton) from fluree-db-binary-index to
fluree-db-core::disk_cache, gated behind the existing native feature, so it can
be shared by other on-disk caches (e.g. Iceberg) under one global budget. The
ContentStore convenience fns move with it (ContentStore is a core type).
binary-index re-exports the API; behavior is unchanged.
